register component renamed to signup, and now styled with all relevant classes from components.html
