+++ src/components/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import Home from './Home';
 import Login from './AuthForms/Login';
-import Register from './AuthForms/Register';
+import SignUp from './AuthForms/SignUp';
 
 function App() {
   return (
@@ -12,7 +12,7 @@ function App() {
           <Route exact path="/" component={Home} />
           <Route exact path="/home" component={Home} />
           <Route path="/home/login" component={Login} />
-          <Route path="/home/register" component={Register} />
+          <Route path="/home/signup" component={SignUp} />
         </Switch>
       </div>
     </BrowserRouter>
